moved retrofit impl to avoid jacoco issues.
